Made buildout-versions buildout 2.0+ compatible.
The presence of an always_unzip argement to _get_dist is checked.
Pre 2.0 buildouts always pass this as False or True. If we get it,
we pass it along. Otherwise we don't.
Turning it into a keyword arg also fixes the '4 instead of 3 args'
error you'd get with buildout 2.0: we monkeypatch the _get_dist.

Also a failing is_distribute import is try/excepted end set to True:
buildout 2.0 always uses/expects distribute instead of setuptools.