Added support for buildout 2.0 greater-than version requirements.
A conditional import that works in buildout 2.0:
from zc.buildout.easy_install import _constrained_requirement
This one is used as it supports '>=', this is mostly only used by buildout internally
to prevent updates from 1.7 to 2.0 and downgrades from 2.0 to 1.7.

This probably means this fix is also needed to get it working in 1.7, but I
haven't checked that.